stdlib: remove unneeded macros from compiled image.
A few macros in the library are local; they are needed only
inside functions in the same module. This means they are not
referenced by anything once those functions are compiled; we
should not be carrying the compiled versions of these macros
into the image. We wrap these macro definitions with
eval-only, telling the compiler not to emit their compiled
version into the output file.

* share/txr/stdlib/asm.tl (with-lev-idx, defopcode,
defopcode-derived): Wrap macro with eval-only.

* share/txr/stdlib/compiler.tl (compile-in-toplevel):
Likewise.

* share/txr/stdlib/path-test.tl (sys:path-test): Likewise.