path-tests: bug in path-strictly-private-to-me-p.
* share/txr/stdlib/path-test.tl (path-strictly-private-to-me):
Fix mistake in test whether the file is group
readable/writable. The wrong flag is used for group
readability test, which means that, at that point in the
logic, if the file is merely not group writable, we jump to
the conclusion that the file is strictly private, when in fact
it may be group readable. Reported by user vapnik spaknik.