open-process: close-on-exec bugfix.
This is a regression introduce in the 2019-06-09 commit
4f33b169dc547b7f9af6f2f1c173d36fc4d62fe8 "Adding fcntl
interface." which changed HAVE_FCNTL_H to HAVE_FCNTL
in the configure script, without making the same change
in stream.c.

* stream.c: Replace HAVE_FCNTL_H with the correct
HAVE_FCNTL, so that we're once again actually including
the <fcntl.h> header now, and setting the close-on-exec
flag on the pipe in open-process.