open-process: I/O redirection feature.
In the mode-string of open-process, I/O redirections can now
be specified that are carried out in the child process.

* stream.c (do_parse_mode): Parse the redirection syntax
indicated by >, which has a short and long form.
The redirections are entered into a small table in the mode
structure.
(open_process): In the child process, perform the redirection
specified in the mode structure's redirection table.

* stream.h (STDIO_MODE_NREDIRS): New preprprocessor symbol.
(struct stdio_mode): New array member, redir.

* txr.1: Documented.