eval: bugfix: expansion wrongly always in null env.
The eval function uses a null macro environment for expanding
a form, even when it's given an environment object. This
causes spurious warnings about unbound variables/functions.
For instance:

    (let ((env (make-env '((x . 42)) nil nil)))
      (eval '(+ x x) env)) ;; warning: unbound variable x

To fix this, we have to create a macro version of the
incoming environment and expand with that.

* eval.c (env_to_menv): Take an evaluation environment chain
and convert it to a (flattened) macro environment.
(expand_eval): Take a macro environment parameter and use
that for expanding the form to be evaluated, rather than nil.
(eval_intrinsic): Calculate a macro environment corresponding
to the given evaluation environment. Use that for the
macroexpand call, and also pass it down to expand_eval
to be used for the full expansion.