configure: disable building as PIE.
On some GNU/Linux distributions, the compiler is configured
for PIE (position-independent executables).

This is horrible garbage which causes our "make tests" to run
more than 22% slower (e.g. 8.22 seconds versus 6.69) which is
unacceptable. Worse, this was implicated in a crash in TXR's
exception handling logic.  The committed workaround for that
problem stays in place for users who want to build TXR in the
PIE style anyway.

* configure: detect which, if any, of the known gcc and clang
options are available for disabling PIE and add those to
opt_flags (which turns into OPT_FLAGS in config.make).