ftw: throw exception on failure.
The ftw function just returns -1 if, for instance, given a bad
path (nonexistent or no permissions).  The documentation also
has issues.

* ftw.c (ftw_wrap): If the return value is -1, convert it to a
file-error using errno_to_file_error and throw a diagnostic
exception. This situation can, of course, arise if the
callback function returns -1, in which case it should prepare
a value in errno.

* txr.1: Return value of callbackfn documented better.
Documented exception throwing behavior, and failed termination
if the callback returns -1.