New function: merge-delete-package.
This is a useful function which supports the use of temporary
packages over the scope of file compilation units.
A file can be read under a temporary package which provides
usefully customized symbol visibility consisting of
an arrangement of symbols from various other packages.
Then, in a single operation, thanks to this new function,
that packag can be deleted and all of its local symbols
(those having been newly interned over the course of the
file) are transferred to some other, more permanent package.

* eval.c (eval_init): merge-delete-package intrinsic
registered.

* lib.c (merge_delete_package): New function.

* lib.h (merge_delete_package): Declared.

* txr.1: Documented.