repl: catch exceptions during completion.
This fixes the issue that TXR exits if an exception occurs
during Tab completion in the interactive listener.
This could happen when loading a file fails for whatever
reason, such as a corrupt or incomplete installation
of the library files or whatever.

* parser.c (provide_completions): Set up a catch all handler
here around everything, like we already did in provide_atom.