chmod: bug handling comma after right hand ugo.
* sysif.c (chmod_wrap): The chm_comma state is transitioned to
after seeing a right hand side u, g or o. These do not combine
with other letters, so ch_comma expects a comma after which a
new permission clause we start,.  Therefore the srcm and who
variables must be rest.  It's also a good idea to continue the
loop.

* tests/018/chmod.tl: New test case which exposed
the above issue.