chmod tests: avoid sticky bit when not available.
On Solaris, we can't set the sticky bit on a non-directory
without special privilege.

Let's detect whether we can set the sticky bit on our test
object. If we can't, then we avoid executing tests that
involve the sticky bit.

* tests/018/chmod.tl (test-sticky): New variable.
(cht): If test-sticky is false, only run the test if none of
the inputs contain a 't'.