@(line): get meaningful value in horizontal mode.
When @(line var) is used in a horizontal match, line is bound
to zero. This is because the line number isn't being
propagated.

* match.c (mf_from_ml): Both the file and line context
structurs have a data_lineno, yet this function doesn't
propagate it. Instead of using ml_all, whose name
tells a subtle lie since it neglects the data_lineno,
let's just do the member for member logic here.
This function is used in one place, the hv_trampoline.
So this fixes the line number for all vertical directives
called out of horizontal context.